+++ src/models/Settings.php
@@ -6,12 +6,21 @@ use craft\base\Model;
 
 class Settings extends Model
 {
-    public $categories = ['site'];
+    public $pluginName = 'Translations';
+    public $categories = [['category' => 'site']];
     public $addMissingTranslations = true;
     public $addMissingSiteRequestOnly = true;
 
     public function getCategories()
     {
-        return $this->categories;
+        $cats = [];
+        foreach ($this->categories as $cat) {
+            if (isset($cat['category'])) {
+                $cats[] = $cat['category'];
+            } else {
+                $cats[] = $cat;
+            }
+        }
+        return $cats;
     }
 }